Refactor: move FloatingText to `src` module
